# swagger/BlueCycle_API_Doc.yaml
openapi: 3.0.0
info:
  title: Swagger BlueCycle - OpenAPI 3.0
  description: |-
    The BlueCycle API provides a comprehensive interface for managing user access, coupon acquisitions, and machine operations. It is designed to serve two primary roles:
    
    - **Users (citizens)**: Enables regular users to access and interact with coupons and manage their personal accounts.
    - **Administrators**: Grants advanced capabilities for effective system management, including interaction with system assets such as coupons and machines.
    
    ### Features and Functionalities
    
    #### User (Citizen) Actions
    Using the BlueCycle API, users can:
    - Authenticate their autentity by providing their login credentials.
    - Register a new account and manage their login credentials.
    - Retrieve personal details, such as login information and points balance.
    - Change their access credentials.
    - List all coupons acquired through the store and add new ones.
    - Retrieve all wishlisted coupons, add new ones and remove any wishlisted item from thei account.
    - Browse the store's available coupons.
    - Acquire new coupons and view their acquired coupons.
    - Review detailed records of their transaction history.
    - List and filter available machines nearby and fetch information for a specific one. 
    
    #### Administrative (Admin-Only) Actions
...
    - Add or remove coupons from the store.
...
    - Manage coupon listed on the store

...
servers:
...
paths:
  /userAuth:
    post:
      tags:
      - user actions
      summary: Allows user authentication (token required)
      description: Authenticates a user using their credentials and returns a JSON Web Token (JWT) for accessing protected endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                  description: The registered user's email address
                password:
                  type: string
                  format: password
                  description: The registered user's password
      responses:
        "200":
          description: Authenticated successfully. Returns a JWT token for authorization
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Authenticated successfully
                  data:
                    type: object
                    properties:
                      self:
                        type: string
                        description: User ID
                      token:
                        type: string
                        description: A JWT token to be used for accessing protected API endpoints
        "400":
          description: Invalid or missing credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Bad request
                  data:
                    example: null
        "404":
          description: No user matches with the provided email
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Not found
                  data:
                    example: null
        "500":
          $ref: "#/components/responses/ServerError"
  /adminAuth:
    post:
      tags:
      - admin actions
      summary: Allows admin authentication (NO AUTH required)
      description: Authenticates an admin using their credentials and returns a JSON Web Token (JWT) for accessing protected endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: admin@bluecycle.com
                  description: The registered admin's email address
                password:
                  type: string
                  format: password
                  description: The registered admin's password
      responses:
        "200":
          description: Authenticated successfully. Returns a JWT token for authorization
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Authenticated successfully
                  data:
                    type: object
                    properties:
                      self:
                        type: string
                        description: Admin ID
                      token:
                        type: string
                        description: A JWT token to be used for accessing protected API endpoints
        "400":
          description: Invalid or missing credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Bad request
                  data:
                    example: null
        "404":
          description: No user matches with the provided email
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Not found
                  data:
                    example: null
        "500":
          $ref: "#/components/responses/ServerError"
  /users:
    get:
      security:
      - AdminAuth: []
      - MachineAuth: []
      tags:
        - admin actions
        - machine actions
      summary: Retrieve a List of all users
      description: Fetches a comprehensive list of users registered within the system. Optional query parameters can be used to filter the results
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: false
          description: Fetches a item by a specific ID among a list of users
        - in: query
          name: email
          schema:
            type: string
          required: false
          description: Filter the user list by a specific email address
        - in: query
          name: code
          schema:
            type: string
          required: false
          description: Fetches a item by a specific code among a list of users
      responses:
        "200":
          description: Successful response containing the list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: OK
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/user"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      tags:
      - user actions
      summary: Adds a new user to the system (no token required)
      description: Creates a new (user) account within the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - name
              properties:
                email:
                  type: string
                  example: new_user@example.com
                name:
                  type: string
                  example: Mario Rossi
        description: Credentials provided by SpID and CieID authentication systems
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: New user registered
                  data:
                    $ref: "#/components/schemas/user"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: User already exists
                  data:
                    example: null
        "400":
          description: Missing body parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Bad request
                  data:
                    example: null
        "500":
          $ref: "#/components/responses/ServerError"
...
  /machines:
    get:
      security:
      - UserAuth: []
      - AdminAuth: []
      tags:
      - admin actions
      - user actions
      summary: Returns a list of machines
      description: Fetches a list of all machines installed and registered in the system
      parameters:
      - in: query
        name: available
        required: false
        schema:
          type: boolean
      - in: query
        name: proximity # vicinanza,...
        description: Returns all machines within [range]km from the given position
        required: false
        schema:
          type: object
          properties:
            range:
              type: integer
              description: In kilometers
            from:
              type: object
              properties:
                latitude:
                  type: number
                  format: double
                longitude:
                  type: number
                  format: double
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: OK
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/machine"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      security:
      - AdminAuth: []
      tags:
      - admin actions
      summary: Adds a new machine
      description: Creates a new machine within the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - position
              - available
              properties:
                position:
                  type: object
                  properties:
                    latitude:
                      type: number
                      format: double
                    longitude:
                      type: number
                      format: double
                available:
                  type: boolean
                description:
                  type: string    
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Created
                  data:
                    $ref: "#/components/schemas/machine"
        "400":
          description: Missing required machine parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Bad request
                  data:
                    example: null
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /couponPrototypes:
    get:
      security:
      - AdminAuth: []
      - UserAuth: []
      tags:
      - admin actions
      - user actions
      summary: Returns a list of coupons
      description: Fetches a list of all coupons available on the system
      parameters:
      - in: query
        name: store
        schema:
          type: string
        required: false
      - in: query
        name: price
        schema:
          type: integer
        required: false
      - in: query
        name: discount
        schema:
          type: integer
        required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: OK
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/couponPrototype"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      security:
      - AdminAuth: []
      tags:
      - admin actions
      summary: Creates a new coupon
      description: Adds a new coupon to the rewards store
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
              - store
              - discount
              - price
              properties:
                store:
                  type: string
                discount:
                  type: integer # coupon pertentage discount
                  minimum: 1
                  maximum: 50
                price:
                  type: integer
                  minimum: 0
                description:
                  type: string
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Created
                  data:
                    $ref: "#/components/schemas/couponPrototype"
        "400":
          description: Missing required parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Bad request
                  data:
                    example: null
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /transactions:
    post:
      security:
      - MachineAuth: []
      tags:
      - machine actions
      summary: Registers a new transacion
      description: Registers a new transaction involving a user and a machine and updates the user's points balance(if the daily quota hasn't been reached)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                  description: Link to the user
                machine:
                  type: string
                  description: Link to the machine
                date:
                  type: string
                  format: date
                  description: In Unix timestamp
                collected:
                  type: integer
      responses:
        "201":
          description: New transaction registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Created
                  data:
                    $ref: "#/components/schemas/transaction"
        "400":
          description: Malformed request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: Not valid id/s
                  data:
                    example: null
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/ServerError"
components:
  schemas:
    user:
      description: Regular user
      type: object
      required:
      - self
      - code
      - email
      - name
      - password
      properties:
        self:
          type: string
          description: The path where the user is located
        code:
          type: string
          description: The unique code used for synchronization at the machine 
        email:
          type: string
          description: The user's email address
          example: user@example.com
        name:
          type: string
          description: The full name of the user
          example: Mario Rossi
        password:
          type: string
          description: The user's password
        points:
          type: integer
          description: The user's points balance
          default: 0
          minimum: 0
          example: 0
...
    machine:
      type: object
      required:
      - self
      - position
      - available
      properties:
        self:
          type: string
          description: The path where the machine is located within the system
        position:
          description: The position where the machine is installed
          type: object
          required:
          - latitude
          - longitude
          properties:
            latitude:
              type: number
              format: double
              description: The latitude of the machine's location
              example: 46.071720
            longitude:
              type: number
              format: double
              description: The longitude of the machine's location
              example: 11.120600
        available:
          type: boolean
          description: The machine availability status
          example: true
        description:
          type: string
...
    couponPrototype:
      description: This is a item available on the store. It represents a coupon pattern
      type: object
      required:
      - self
      - store
      - discount
      - price
      properties:
        self:
          type: string
          description: The path where the coupon prototype is located within the system
        store:
          type: string
          description: The name of the store offering the coupon
        discount:
          type: integer
          description: The percentage discount offered by the coupon
          minimum: 1
          maximum: 100 #to be discussed
          example: 15
        price:
          type: integer
          description: The amount of points required to purchase the coupon
        description:
          type: string
          description: A description of the coupon
...
    transaction:
      type: object
      required:
      - self
      - machine
      - user
      - date
      - collected
      properties:
        self:
          type: string
          description: The path where the transaction is located within the system
        user:
          type: string
          description: The link to the user involved in the transaction
        machine:
          type: string
          description: The link to the machine involved in the transaction
        date:
          type: integer
          description: The date of the transaction in Unix timestamp
          example: 1732557477
        collected:
          type: integer
          description: The number of bottles collected. Each collected bottle is a point added to the user's profile
          example: 4
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication scheme for administrators using JSON Web Tokens (JWT)
    UserAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication scheme for regular users using JSON Web Tokens (JWT)
    MachineAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication scheme for machines using JSON Web Tokens (JWT)
  responses:
...
    UnauthorizedError:  #401
      description: Unauthorized access. Authentication credentials are missing or invalid(e.g. expired token)
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Unauthorized
              data:
                example: null
    ForbiddenError:   #403
      description: Forbidden access. The authenticated user does not have the required permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Forbidden
              data:
                example: null
    ServerError:  #500
      description: Something went wrong while processing request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Internal server error
              data:
                example: null